ENH: Update MLFG to 64 bits
Replace 32 bit implementation of MLFG wiht a 64 bit one, doubling performance
for random doubles
Finish doc strings

+++ randomstate/src/mlfg-1279-861/mlfg-1279-861-test-gen.c
@@ -1,12 +1,17 @@
+/*
+*  gcc -O2 ../splitmix64/splitmix64.c mlfg-1279-861.c mlfg-1279-861-test-gen.c -o mlfg -std=c99
+*/
+
 #include <stdio.h>
 #include <inttypes.h>
+#include "../splitmix64/splitmix64.h"
 #include "mlfg-1279-861.h"
 
 int main(void)
 {
     int i;
     uint64_t seed = 1ULL;
-    uint32_t temp;
+    uint64_t temp;
     mlfg_state state = {{ 0 }};
     mlfg_seed(&state, seed);
 
@@ -20,8 +25,8 @@ int main(void)
     for (i=0; i < 1000; i++)
     {
         temp = mlfg_next(&state);
-        fprintf(fp, "%d, %" PRIu32 "\n", i, temp);
-        printf("%d, %" PRIu32 "\n", i, temp);
+        fprintf(fp, "%d, %" PRIu64 "\n", i, temp);
+        printf("%d, %" PRIu64 "\n", i, temp);
     }
     fclose(fp);
 
@@ -36,8 +41,8 @@ int main(void)
     for (i=0; i < 1000; i++)
     {
         temp = mlfg_next(&state);
-        fprintf(fp, "%d, %" PRIu32 "\n", i, temp);
-        printf("%d, %" PRIu32 "\n", i, temp);
+        fprintf(fp, "%d, %" PRIu64 "\n", i, temp);
+        printf("%d, %" PRIu64 "\n", i, temp);
     }
     fclose(fp);
 }
